feat(fav5): created compatible iconstack endpoint
you can use the url format from v1 and swap to v2 while preserving the general layout
